Keep the same ordering of the PATH directories when preprending the virtual environment directory
This ensures that the directories within the virtual environment are
called in the same order they would be outside. Before, it would put
directories at the end of the original PATH at the beginning. However,
we want to keep the order so the virtual environment works the same as
not using the environment.