Only install filesystem instead of pacman during environment creation
It isn't necessary to install Pacman during the initial creation of the
virtual environment. This will most likely be installed pretty quickly
once new packages get created, but only 'filesystem' and 'sed' are
necessary for the virtual environment to work.

This change speeds up virtual environment creation and reduces initial
size because pacman (and therefore its dependencies) don't need to be
installed.