Add the ability to run pacman without root privileges in a virtual environment
This is done using the unshare command. It works with both pacman and
yay currently. This could definitely be a very attractive feature for
pacman-venv in the future if done right.

I want to work on overlaying the pacman cache as a read-only directory
so users don't download packages they already have installed, saving
space and time.